bind: Split options from bound_contract.go
